+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Added
+
+- Add `:full_page` option to `ChromicPDF.capture_screenshot/2` to increase the viewport size to match the content.
+
 ## [1.9.1] - 2024-06-08
 
 ### Fixed